Force use of the stuf readline implementation
Term::ReadLine is an umbrella module which chooses which underlying
implementation to use based on what is installed.  Unfortunately, they
are not all alike; Term::ReadLine::Perl, for instance, inserts an extra
blank line (filled to the terminal width with spaces) before the prompt.
This causes t/simple.t to fail.

This would usually be addressed by switching to a regular expression to
sop up the extra spaces.  However, the entire point of the failing test
in t/simple.t is to test `expect_is`; `expect_like` is tested
separately.  Thus, the test _demands_ a reliable output from
Term::ReadLine.

Moreover, the Test::Expect test is not the only one that would expect
that output be reliable from deployment-to-deployment; tests which use
Test::Expect will likely similarly expect consistency in the programs
they test.

To this end, enforce (using the PERL_RL environment variable) the "Stub"
implementation of Term::Readline upon programs called by Test::Expect.
The stub is guaranteed to be installed if Term::ReadLine is, offers all
of the functionality that Test::Expect needs (the ability to read
lines), and produces consistent, reliable output.